moved code to module
