+++ docs/new_spec.txt
