+++ src/module/__init__.py
@@ -1,0 +1,1 @@
+from .game import Card, District, Game, Player, Role, load_game_data
